Correct warnings given by gometalinter
The gometalinter tool reported errors.

Cyclomatic complexity in the Marshal/Unmarshal functions
and also in the test function Test_Types.
Use extract method to split up the functions in logical parts.

Error return values not being checked in the Marshal/Unmarshal
and in the NewConfigFile and ConfigFile#Persist functions.
Check these values and return errors if they occur.

This gives a change in the internal logic of Marshal/Unmarshal,
NewConfigFile and ConfigFile#Persist.
The previous implementation ignored errors when writing/setting values,
writing to file and closing the file handle.
The new implementation will abort and report these errors, and
the function will fail.